2.1 Per-step int8 clamp (parse + USB)

+++ usb-output/main/macpass_hid.h
@@ -8,7 +8,7 @@ void hid_wake_pump(void);
 void hid_add_report(hid_transmit_t report);
 
 /** Queue a macro step delta; drips out on a timer and on SPI mouse reports (rate-limited). */
-void hid_macro_feed_mouse_step(int8_t x, int8_t y, int8_t wheel, int8_t pan, uint32_t step_us);
+void hid_macro_feed_mouse_step(int16_t x, int16_t y, int16_t wheel, int16_t pan, uint32_t step_us);
 void hid_macro_cancel_mouse_spread(void);
 bool hid_macro_mouse_spread_active(void);
 /** Restart periodic drip timer using `macro_profile_mouse_drip_interval_us()`. */
